# quikthread-backend/openapi.yaml
openapi: 3.0.3
info:
  title: QuikThread API
  description: |
    Python FastAPI backend for QuikThread - Audio/Video to X Thread Generator
    
    This API allows users to:
    - Upload audio/video files or provide URLs (YouTube, TikTok, Instagram, Twitter, etc.)
    - Transcribe media using Deepgram
    - Generate viral X (Twitter) threads using Google Gemini AI
    - Post threads directly to X/Twitter
    - View analytics for posted threads
    
    ## Authentication
    All endpoints (except webhooks and health checks) require Firebase Authentication.
    Include the Firebase ID token in the Authorization header:
    ```
    Authorization: Bearer <firebase-id-token>
    ```
    
    ## Rate Limiting
    API usage is limited by user tier:
    - **Free**: 2 credits/month, 30-minute max duration
    - **Pro**: 50 credits/month, 60-minute max duration, post to X
    - **Business**: Unlimited credits, 60-minute max duration, post to X, analytics
  version: 1.0.0
  contact:
    name: QuikThread Support
    url: https://quikthread.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.quikthread.com
    description: Production server
...
components:
  securitySchemes:
    FirebaseAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase ID token obtained from Firebase Authentication
  schemas:
    Error:
      type: object
      properties:
        detail:
          oneOf:
            - type: string
            - type: object
          description: Error message or detailed error object
      required:
        - detail
...
    ProcessRequest:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum: [upload, url]
          description: Type of input - 'upload' for Firebase Storage URL or 'url' for direct URL
        fileUrl:
          type: string
          format: uri
          description: Firebase Storage URL for uploaded files (required if type=upload)
        contentUrl:
          type: string
          format: uri
          description: Direct URL to audio/video content (required if type=url). Supports YouTube, TikTok, Instagram, Twitter, Facebook, Vimeo, Twitch, Spotify, and podcasts.
    JobStatus:
      type: object
      properties:
        jobId:
          type: string
          description: Unique job identifier
        status:
          type: string
          enum: [processing, transcribing, generating, completed, failed]
          description: Current job status
        progress:
          type: integer
          minimum: 0
          maximum: 100
          description: Progress percentage (0-100)
        type:
          type: string
          description: Type of job - 'upload' or 'url'
        duration:
          type: number
          format: float
          nullable: true
          description: Audio/video duration in seconds
        threads:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/Thread'
          description: Generated thread options
        error:
          type: string
          nullable: true
          description: Error message if job failed
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
...
    Thread:
      type: object
      properties:
        threadNumber:
          type: integer
          description: Thread option number (1-5)
        hook:
          type: string
          description: Description of the hook strategy used
        tweets:
          type: array
          items:
            type: string
          description: Array of tweet texts that form the thread
...
paths:
...
  /api/process:
    post:
      tags:
        - Process
      summary: Process audio/video
      description: |
        Create a processing job to transcribe audio/video and generate X threads.
        
        **Process:**
        1. Extract media URL (if social media URL)
        2. Transcribe using Deepgram
        3. Generate threads using Gemini AI
        
        **Quota:** Automatically checks and deducts 1 credit.
        
        **Polling:** Use GET /api/jobs/{jobId} to check status.
      security:
        - FirebaseAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProcessRequest'
      responses:
        '202':
          description: Job created and processing started
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  jobId:
                    type: string
                    description: Job ID for tracking progress
                  message:
                    type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Quota exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to create job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/jobs/{job_id}:
    get:
      tags:
        - Process
      summary: Get job status
      description: Get the current status and results of a processing job
      security:
        - FirebaseAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
          description: Job ID returned from POST /api/process
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Access denied - job belongs to another user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to retrieve job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/jobs:
    get:
      tags:
        - Process
      summary: Get user jobs
      description: Get all jobs for the current user (last 50)
      security:
        - FirebaseAuth: []
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/JobStatus'
                  count:
                    type: integer
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to retrieve jobs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...